+++ core/modules/modsplit.class.php
@@ -64,7 +64,7 @@ class modsplit extends DolibarrModules
 		$this->description = "Description of module split";
 		// Possible values for version are: 'development', 'experimental' or version
 
-		$this->version = '2.3.0';
+		$this->version = '2.3.1';
 
 		// Key used in llx_const table to save module status enabled/disabled
 		// (where MYMODULE is value of property name of module in uppercase)
@@ -117,9 +117,9 @@ class modsplit extends DolibarrModules
 		// List of modules id to disable if this one is disabled
 		$this->requiredby = array();
 		// Minimum version of PHP required by module
-		$this->phpmin = array(5, 3);
+		$this->phpmin = array(7, 0);
 		// Minimum version of Dolibarr required by module
-		$this->need_dolibarr_version = array(3, 2);
+		$this->need_dolibarr_version = array(15, 0);
 		$this->langfiles = array("split@split"); // langfiles@split
 		// Constants
 		// List of particular constants to add when module is enabled