+++ ChangeLog.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 - NEW : change user->right to hasRight() function - *23/11/2023* - 2.3.0  
+      $conf->global to getDolGlobal
+
 
 ## Version 2.2