+++ ChangeLog.md
@@ -4,7 +4,8 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 - NEW : change user->right to hasRight() function - *23/11/2023* - 2.3.0  
-      $conf->global to getDolGlobal
+        $conf->global to getDolGlobal
+        add commonhookActions
 
 
 ## Version 2.2