+++ ChangeLog.md
@@ -3,7 +3,11 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
-- NEW : add min php/dolibarr version  - *30/11/2023* - 2.3.1  
+
+
+## Release 2.4
+
+- NEW : add min php/dolibarr version  - *30/11/2023* - 2.4.1  
 - NEW : change user->right to hasRight() function - *23/11/2023* - 2.3.0  
         $conf->global to getDolGlobal
         add commonhookActions
@@ -12,7 +16,7 @@ All notable changes to this project will be documented in this file.
         trad error missing  :  line selection
 
 
-## Version 2.2
+## Release 2.2
 
 - FIX : Editor name  *03/08/2022* 2.2.3
 - FIX : Compat PHP 8  *02/08/2022* 2.2.2
@@ -20,7 +24,7 @@ All notable changes to this project will be documented in this file.
 - NEW : Ajout de la class TechATM pour l'affichage de la page "A propos" *11/05/2022* 2.2.0
 - NEW : Prise en compte des objets Facture et Commande *12/01/2021* - 2.1.0
 
-## Version 2.0
+## Release 2.0
 - FIX : change filter setting for  forgeSQLFromUniversalSearchCriteria  in $form->select_company on showlines.php - *12/06/2023* - 2.0.4  
 - FIX : V16 FAMILY - *02/06/2022* - 2.0.3  
 - FIX: Compatibility with Dolibarr 14.0 + remove unused files - *21/10/2021* - 2.0.2